feat: fuzzy filter (#528)
